fix errors in data tests

+++ data/break_his/break_his_dataset_creation.py
@@ -8,36 +8,37 @@ import logging
 logging.basicConfig(level=logging.INFO)
 _logger = logging.getLogger(__name__)
 
-current_dir = os.path.dirname(os.path.abspath(__file__))
-
 benign = ['adenosis', 'fibroadenoma', 'phyllodes_tumor', 'tubular_adenoma']
 malignant = ['ductal_carcinoma', 'lobular_carcinoma', 'mucinous_carcinoma', 'papillary_carcinoma']
 
 
-def extract_and_create(archived_dataset):
+def extract_and_create(archived_dataset, current_dir):
     tar = tarfile.open(archived_dataset)
     tar.extractall()
     tar.close()
 
     os.remove(archived_dataset)
-    os.mkdir('break_his_train')
-    os.mkdir('break_his_validation')
-    os.mkdir('break_his_test')
+    os.mkdir(os.path.join(current_dir, 'break_his_train'))
+    os.mkdir(os.path.join(current_dir, 'break_his_validation'))
+    os.mkdir(os.path.join(current_dir, 'break_his_test'))
 
     for directory in ['break_his_train', 'break_his_validation', 'break_his_test']:
         for cancer_type in [benign, malignant]:
             for cancer_subtype in cancer_type:
-                os.mkdir(directory + '/' + cancer_subtype)
+                os.mkdir(os.path.join(current_dir, directory + '/' + cancer_subtype))
 
 
-def move_to_train(cancer_type, cancer_subtypes):
+def move_to_train(cancer_type, cancer_subtypes, current_dir):
     for cancer_subtype in cancer_subtypes:
+        subdirectories = []
         if cancer_type == 'benign':
-            subdirectories = [f.path for f in os.scandir(current_dir + '/BreaKHis_v1/histology_slides/breast/benign/SOB/' +
-                              cancer_subtype) if f.is_dir()]
+            if os.path.isdir(current_dir + '/BreaKHis_v1/histology_slides/breast/benign/SOB/' + cancer_subtype):
+                subdirectories = [f.path for f in os.scandir(current_dir + '/BreaKHis_v1/histology_slides/breast/benign/SOB/' +
+                                  cancer_subtype) if f.is_dir()]
         else:
-            subdirectories = [f.path for f in os.scandir(current_dir + '/BreaKHis_v1/histology_slides/breast/malignant/SOB/' +
-                              cancer_subtype) if f.is_dir()]
+            if os.path.isdir(current_dir + '/BreaKHis_v1/histology_slides/breast/malignant/SOB/' + cancer_subtype):
+                subdirectories = [f.path for f in os.scandir(current_dir + '/BreaKHis_v1/histology_slides/breast/malignant/SOB/' +
+                                  cancer_subtype) if f.is_dir()]
         for subdirectory in subdirectories:
             source_directory = subdirectory + '/100X/'
             destination_directory = current_dir + '/break_his_train/' + cancer_subtype
@@ -46,7 +47,7 @@ def move_to_train(cancer_type, cancer_subtypes):
                 shutil.move(source_directory + image, destination_directory)
 
 
-def move_to_directory(directory, cancer_subtype, total, number_to_move):
+def move_to_directory(directory, cancer_subtype, total, number_to_move, current_dir):
     files = [f.path for f in os.scandir(current_dir + '/break_his_train/' + cancer_subtype)]
     random_list = random.sample(range(0, total-1), number_to_move)
     files_to_move = []
@@ -56,26 +57,23 @@ def move_to_directory(directory, cancer_subtype, total, number_to_move):
         shutil.move(file_path, current_dir + directory + cancer_subtype)
 
 
-def main(archived_dataset):
+def main(archived_dataset, current_dir):
     try:
-        extract_and_create(archived_dataset)
+        extract_and_create(archived_dataset, current_dir)
 
-        move_to_train('benign', benign)
-        move_to_train('malignant', malignant)
+        move_to_train('benign', benign, current_dir)
+        move_to_train('malignant', malignant, current_dir)
 
-        shutil.rmtree(current_dir + '/BreaKHis_v1')
+        if os.path.isdir(current_dir + '/BreaKHis_v1'):
+            shutil.rmtree(current_dir + '/BreaKHis_v1')
 
         for cancer_subtype in [*benign, *malignant]:
-            total = len([name for name in os.listdir(current_dir + '/break_his_train/' + cancer_subtype)])
-            number_to_move = int(total*0.15)
-            move_to_directory('/break_his_validation/', cancer_subtype, total, number_to_move)
-            move_to_directory('/break_his_test/', cancer_subtype, total-number_to_move, number_to_move)
+            if os.path.isdir(current_dir + '/break_his_train/' + cancer_subtype):
+                total = len([name for name in os.listdir(current_dir + '/break_his_train/' + cancer_subtype)])
+                number_to_move = int(total*0.15)
+                move_to_directory('/break_his_validation/', cancer_subtype, total, number_to_move, current_dir)
+                move_to_directory('/break_his_test/', cancer_subtype, total-number_to_move, number_to_move, current_dir)
     except Exception as e:
         _logger.error('Exception caught in main: {}'.format(e), exc_info=True)
         return 1
     return 0
-
-
-if __name__=='__main__':
-    archived_dataset = 'BreaKHis_v1.tar.gz'
-    exit(main(archived_dataset))
