+++ data/nct_crc_he_100k/nct_crc_he_100k_dataset_creation.py
@@ -8,8 +8,6 @@ import logging
 logging.basicConfig(level=logging.INFO)
 _logger = logging.getLogger(__name__)
 
-current_dir = os.path.dirname(os.path.abspath(__file__))
-
 tissue_types = ['adipose', 'background', 'debris', 'lymphocytes', 'mucus', 'smooth_muscle', 'normal_colon_mucosa',
                 'cancer_associated_stroma', 'colorectal_adenocarcinoma_epithelium']
 
@@ -18,23 +16,23 @@ tissue_types_dict = {'adipose': 'ADI', 'background': 'BACK', 'debris': 'DEB', 'l
                      'colorectal_adenocarcinoma_epithelium': 'TUM'}
 
 
-def extract_and_create(archived_dataset):
+def extract_and_create(archived_dataset, current_dir):
     with zipfile.ZipFile(archived_dataset, 'r') as zip_ref:
         zip_ref.extractall(current_dir)
 
     os.remove(archived_dataset)
 
-    os.mkdir('nct_crc_he_100k_train')
-    os.mkdir('nct_crc_he_100k_validation')
-    os.mkdir('nct_crc_he_100k_test')
+    os.mkdir(os.path.join(current_dir, 'nct_crc_he_100k_train'))
+    os.mkdir(os.path.join(current_dir, 'nct_crc_he_100k_validation'))
+    os.mkdir(os.path.join(current_dir, 'nct_crc_he_100k_test'))
 
     for directory in ['nct_crc_he_100k_train', 'nct_crc_he_100k_validation', 'nct_crc_he_100k_test']:
         for tissue_type in tissue_types:
-            os.mkdir(directory + '/' + tissue_type)
+            os.mkdir(os.path.join(current_dir, directory + '/' + tissue_type))
 
 
 
-def move_to_train():
+def move_to_train(current_dir):
     for tissue_type in tissue_types:
         source_directory = current_dir + '/NCT-CRC-HE-100K/' + tissue_types_dict[tissue_type] + '/'
         destination_directory = current_dir + '/nct_crc_he_100k_train/' + tissue_type
@@ -44,7 +42,7 @@ def move_to_train():
                 shutil.move(source_directory + image, destination_directory)
 
 
-def move_to_directory(directory, tissue_type, total, number_to_move):
+def move_to_directory(directory, tissue_type, total, number_to_move, current_dir):
     files = [f.path for f in os.scandir(current_dir + '/nct_crc_he_100k_train/' + tissue_type)]
     random_list = random.sample(range(0, total-1), number_to_move)
     files_to_move = []
@@ -54,24 +52,21 @@ def move_to_directory(directory, tissue_type, total, number_to_move):
         shutil.move(file_path, current_dir + directory + tissue_type)
 
 
-def main(archived_dataset):
+def main(archived_dataset, current_dir):
     try:
-        extract_and_create(archived_dataset)
-        move_to_train()
+        extract_and_create(archived_dataset, current_dir)
+        move_to_train(current_dir)
 
-        shutil.rmtree(current_dir + '/NCT-CRC-HE-100K')
+        if os.path.isdir(current_dir + '/NCT-CRC-HE-100K'):
+            shutil.rmtree(current_dir + '/NCT-CRC-HE-100K')
 
         for tissue_type in tissue_types:
-            total = len([name for name in os.listdir(current_dir + '/nct_crc_he_100k_train/' + tissue_type)])
-            number_to_move = int(total*0.15)
-            move_to_directory('/nct_crc_he_100k_validation/', tissue_type, total, number_to_move)
-            move_to_directory('/nct_crc_he_100k_test/', tissue_type, total-number_to_move, number_to_move)
+            if os.path.isdir(current_dir + '/nct_crc_he_100k_train/' + tissue_type):
+                total = len([name for name in os.listdir(current_dir + '/nct_crc_he_100k_train/' + tissue_type)])
+                number_to_move = int(total*0.15)
+                move_to_directory('/nct_crc_he_100k_validation/', tissue_type, total, number_to_move, current_dir)
+                move_to_directory('/nct_crc_he_100k_test/', tissue_type, total-number_to_move, number_to_move, current_dir)
     except Exception as e:
         _logger.error('Exception caught in main: {}'.format(e), exc_info=True)
         return 1
     return 0
-
-
-if __name__=='__main__':
-    archived_dataset = 'NCT-CRC-HE-100K.zip'
-    exit(main(archived_dataset))
